Move ecl::powf to utils.hpp
