+++ EKF/common.h
@@ -43,23 +43,6 @@
 
 #include <matrix/math.hpp>
 
-namespace ecl{
-	inline float powf(float x, int exp)
-	{
-		float ret;
-		if (exp > 0) {
-			ret = x;
-			for (int count = 1; count < exp; count++) {
-				ret *= x;
-			}
-			return ret;
-		} else if (exp < 0) {
-			return 1.0f / ecl::powf(x, -exp);
-		}
-		return 1.0f;
-	}
-}
-
 namespace estimator
 {
