+++ EKF/utils.hpp
@@ -1,5 +1,6 @@
 #include <matrix/math.hpp>
 
+#pragma once
 
 // converts Tait-Bryan 312 sequence of rotations from frame 1 to frame 2
 // to the corresponding rotation matrix that rotates from frame 2 to frame 1
@@ -18,3 +19,20 @@ float kahanSummation(float sum_previous, float input, float &accumulator);
 // calculate the inverse rotation matrix from a quaternion rotation
 // this produces the inverse rotation to that produced by the math library quaternion to Dcmf operator
 matrix::Dcmf quatToInverseRotMat(const matrix::Quatf &quat);
+
+namespace ecl{
+	inline float powf(float x, int exp)
+	{
+		float ret;
+		if (exp > 0) {
+			ret = x;
+			for (int count = 1; count < exp; count++) {
+				ret *= x;
+			}
+			return ret;
+		} else if (exp < 0) {
+			return 1.0f / ecl::powf(x, -exp);
+		}
+		return 1.0f;
+	}
+}
